Check for null resourceName before attempted to delete a cert
Without this, the sdk will send a null resourceName to the server,
resulting in a somewhat misleading 404 error being returned.

+++ src/project-management/android-app.ts
@@ -103,6 +103,12 @@ export class AndroidApp {
   }
 
   public deleteShaCertificate(certificateToDelete: ShaCertificate): Promise<void> {
+    if (!certificateToDelete.resourceName) {
+      throw new FirebaseProjectManagementError(
+          'invalid-argument',
+          'Specified certificate does not include a resourceName. (Use AndroidApp.getShaCertificates() to retrieve ' +
+              'certificates with a resourceName.');
+    }
     return this.requestHandler.deleteResource(certificateToDelete.resourceName);
   }
